refactor(typings): modify  ISchemaFormProps、IFormButtonGroupProps typings
