+++ packages/antd/src/type.tsx
@@ -36,12 +36,10 @@ export interface ISchemaFormProps<V = any> {
   initialValues?: V
   defaultValue?: V
   value?: V
-  editable?: (name: string) => boolean | boolean
+  editable?: boolean | ((name: string) => boolean)
   effects?: IEffects
   locale?: {
-    [key: string]: {
-      [k in string]: string | number
-    }
+    [k in string]: any
   }
   schema?: ISchema
   onChange?: (values: V) => void
@@ -126,7 +124,6 @@ export interface IFormButtonGroupProps {
   className?: string
 
   triggerDistance?: number
-  offsetDistance?: any
   zIndex?: number
   span?: number
   offset?: number
